Changed WeatherStation.py to avoid USB activity when weather station console is believed to be writing to memory.
Note: this is an experimental change that might have some side effects
and might not cure the USB hangup problem. In particular, it polls the
station four times as often when waiting for the data to change, so if
it's got the clock time wrong it's more likely to trigger USB hangup!