Split 'weather.ini' into config and status.
There are now two files, 'weather.ini' and 'status.ini'. The former is
used for configuration stuff and the latter to record status such as
last update times. It should not need to be edited in normal use.

This is quite a substantial change, and may have broken something I
haven't found in my testing. Let me know if there are any problems.

The documentation has yet to be updated.