Added ability to set 'flags' for templates.
The only flag recognised so far is 'L', for 'copy to local directory
instead of uploading'. It can be applied to text or graph templates. To
set a flag, replace the template file name in a task list with a tuple
of file name and flags. For example:

text = ['24hrs.txt', '6hrs.txt', ('yowindow.xml', 'L')]