Tweak to 'live log' sync algorithm.
My new weather station was losing sync when it loses contact with the
outside sensors. The data changes at unexpected times, so I was wrongly
setting the 'sensor clock' which led to a USB lockup. The sync algorithm
now ignores unexpected changes if there's no contact with the outside
sensors.