Added 'low_battery' status.
I've recently found out that the outside sensors send a 'low battery'
warning by setting bit 7 of the 'wind direction' byte. I've modified
pywws/WeatherStation.py to tack this on to the existing 'status' value
as bit 11.
On my weather station this status seems to be set for short periods
quite often. I think it's only meaningful if it stays on for days.