Improved hourly summary when missing data.
When contact is lost with the outside sensors for an extended period the
hourly summary data would include readings from around xx:45. Now it
uses the last timestamp in the hour (e.g. xx:58) just as it does when
data is not missing, unless some data was received in the last 15
minutes of the hour.