Moved wind velocity filter into a separate class.
This should give the same results as previously, but they are stored as
float rather than int. Reprocessing your old data is optional.
The new filter class also has a 'decay' option to make a filter that
gives more weight to the most recent values. This can be used in a user
calibration class to filter wind direction.