Added candlesticks plot as per my own request in issue #42
You can now specify tuples like ``<ycalc>(1, 2, 3)</ycalc>``. If the
value isn't a tuple, it gets converted to one. The lines for the gnuplot
data file now adapt to the number of elements in the tuple.

Two new styles: ``<style>candlesticks 1.5</style>`` is a candlestick graph
with a line width of 1.5. ``<style>candlesticksw 2 0.5</style>`` is a
candlesticks graph with a line width of 2 and whiskerbars with a width
of 50%.

I also added an example xml. All previous functionality shouldn't be
affected.

Signed-off-by: Markus Birth <mbirth@gmail.com>