Allow sync to station that's not logging data
A possible fault mode on some stations is that they stop logging data.
This upsets the logging initialisation that attempts to find the time
stamp of the last logged data, as discussed in bug report #81. This fix
reverts to the quick and inaccurate sync mode when this happens.

Solar stations can also temporarily stop logging if the solar data (sent
at one minute intervals) arrives at the expected logging time.

Signed-off-by: Jim Easterbrook <jim@jim-easterbrook.me.uk>